ai_market_analyzer/requirements.txt
# Binance API và WebSocket
python-binance==1.0.19
websocket-client==1.7.0

# Machine Learning & Data Processing
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.2
tensorflow==2.15.0
keras==2.15.0

# Technical Analysis
ta==0.11.0
pandas-ta==0.3.14b0

# API Framework
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3

# Database
redis==5.0.1
pymongo==4.6.1

# Utilities
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.1
schedule==1.2.0

# Logging & Monitoring
loguru==0.7.2

# Data Visualization (optional)
matplotlib==3.7.1
plotly==5.18.0
